chore: update dev tooling
