+++ .changeset/config.json
@@ -1,0 +1,16 @@
+{
+  "$schema": "https://unpkg.com/@changesets/config@2.0.0/schema.json",
+  "changelog": [
+    "@changesets/changelog-github",
+    {
+      "repo": "marko-js/vite"
+    }
+  ],
+  "updateInternalDependencies": "patch",
+  "baseBranch": "main",
+  "access": "public",
+  "commit": false,
+  "linked": [],
+  "ignore": [],
+  "fixed": []
+}
