--- .github/workflows/release.yml
@@ -1,27 +1,0 @@
-name: Release
-
-on:
-  workflow_run:
-    branches: ["main"]
-    workflows: ["CI"]
-    types: [completed]
-
-jobs:
-  release:
-    if: ${{ github.event.workflow_run.conclusion == 'success' }}
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-      - name: Setup node
-        uses: actions/setup-node@v2
-        with:
-          node-version: 16.x
-          cache: "npm"
-      - name: Install dependencies
-        run: npm ci
-      - name: Release
-        run: npm run release
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
