+++ .mocharc.json
@@ -2,7 +2,6 @@
   "exit": true,
   "timeout": 20000,
   "extension": ["js", "ts", ".marko"],
-  "enable-source-maps": true,
   "watchFiles": ["src/**/*.ts", "src/**/*.marko"],
-  "require": ["esbuild-register", "mocha-snap"]
+  "require": ["tsx", "mocha-snap"]
 }
