--- .releaserc.json
@@ -1,3 +1,0 @@
-{
-  "branches": ["main"]
-}
