+++ package.json
@@ -6,41 +6,40 @@
   "bugs": "https://github.com/marko-js/vite/issues",
   "dependencies": {
     "anymatch": "^3.1.2",
-    "domelementtype": "^2.2.0",
-    "domhandler": "^4.3.0",
-    "htmlparser2": "^7.2.0"
+    "domelementtype": "^2.3.0",
+    "domhandler": "^5.0.3",
+    "htmlparser2": "^8.0.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "^16.2.1",
-    "@commitlint/config-conventional": "^16.2.1",
-    "@marko/compiler": "^5.19.1",
-    "@marko/fixture-snapshots": "^2.1.6",
-    "@marko/testing-library": "^6.0.0",
+    "@changesets/changelog-github": "^0.4.5",
+    "@changesets/cli": "^2.23.0",
+    "@marko/compiler": "^5.21.6",
+    "@marko/fixture-snapshots": "^2.1.7",
+    "@marko/testing-library": "^6.1.2",
     "@types/jsdom": "^16.2.14",
-    "@types/mocha": "^9.1.0",
-    "@types/node": "^17.0.21",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^18.0.0",
     "@types/serve-handler": "^6.1.1",
-    "@typescript-eslint/eslint-plugin": "^5.14.0",
-    "@typescript-eslint/parser": "^5.14.0",
+    "@typescript-eslint/eslint-plugin": "^5.30.0",
+    "@typescript-eslint/parser": "^5.30.0",
     "cross-env": "^7.0.3",
-    "esbuild": "^0.14.25",
-    "esbuild-register": "^3.3.2",
-    "eslint": "^8.10.0",
+    "esbuild": "^0.14.47",
+    "eslint": "^8.18.0",
     "eslint-config-prettier": "^8.5.0",
     "fast-glob": "^3.2.11",
     "fixpack": "^4.0.0",
-    "husky": "^7.0.4",
+    "husky": "^8.0.1",
     "jsdom": "^19.0.0",
-    "lint-staged": "^12.3.5",
-    "marko": "^5.19.0",
-    "mocha": "^9.2.1",
+    "lint-staged": "^13.0.3",
+    "marko": "^5.21.2",
+    "mocha": "^10.0.0",
     "mocha-snap": "^4.3.0",
     "nyc": "^15.1.0",
-    "playwright": "^1.19.2",
-    "prettier": "^2.5.1",
-    "semantic-release": "^19.0.2",
+    "playwright": "^1.23.0",
+    "prettier": "^2.7.1",
     "serve-handler": "^6.1.3",
-    "typescript": "^4.6.2",
+    "tsx": "^3.6.0",
+    "typescript": "^4.7.4",
     "vite": "3.0.0-beta.5"
   },
   "exports": {
@@ -69,27 +68,26 @@
   "module": "./dist/index.mjs",
   "peerDependencies": {
     "@marko/compiler": "^5",
-    "vite": "^2"
+    "vite": "^2 || ^3"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/marko-js/vite"
   },
   "scripts": {
-    "build": "tsc -b && node -r esbuild-register build",
-    "ci:test": "nyc npm run mocha -- --forbid-pending --forbid-only",
+    "build": "tsc -b && tsx build.mts",
+    "ci:test": "nyc npm test -- --forbid-only",
     "format": "npm run lint:eslint -- --fix && npm run lint:prettier -- --write && (fixpack || true)",
     "lint": "tsc -b && npm run lint:eslint && npm run lint:prettier -- -l && fixpack",
     "lint:eslint": "eslint -f visualstudio .",
     "lint:prettier": "prettier \"./**/*{.ts,.js,.json,.md,.yml,rc}\"",
-    "mocha": "cross-env NODE_ENV=test mocha \"./src/**/__tests__/*.test.ts\"",
     "prepare": "husky install",
-    "prepublishOnly": "npm run build",
-    "release": "semantic-release",
+    "release": "npm run build && changeset publish",
     "report": "open ./coverage/lcov-report/index.html",
-    "test": "npm run mocha -- --watch",
+    "test": "cross-env NODE_ENV=test mocha \"./src/**/__tests__/*.test.ts\"",
     "test:inspect": "npm test -- --inspect",
-    "test:update": "npm run mocha -- --update"
+    "test:update": "npm test -- --update",
+    "version": "changeset version && npm i --package-lock-only"
   },
   "types": "dist/index.d.ts"
 }
